Zombies wasnt moving because m_Alive was never being set to true. (fixed) Added some balancings to make it a little harder (Arena max size)

+++ ZombieShooter/include/Player.hpp
@@ -48,7 +48,7 @@ class Player
     void stopUp();
     void stopDown();
 
-    void update(float elapsedTime, Vector2i mousePosition);
+    void update(float elapsedTime, Vector2f mousePosition);
     void upgradeSpeed();
     void upgradeHealth();
     void increaseHealthLevel(int amount);
